pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoresearch"
version = "0.1.0"
description = "Autonomous research loop with a local LLM critic — generalization of karpathy/autoresearch."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "CC-BY-NC-SA-4.0" }
authors = [{ name = "Pedro Carvalho Brom", email = "pcbrom@gmail.com" }]
keywords = ["autonomous-research", "llm-critic", "ollama", "experiment-loop", "hyperparameter-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0",
]

[project.optional-dependencies]
torch  = ["torch>=2.0"]                    # only used to free CUDA cache between iters
xgb    = ["xgboost>=2.0", "scikit-learn>=1.3"]   # for the xgboost_tuning example
tsp    = ["numpy>=1.24"]                          # for the tsp_heuristic example
all    = ["torch>=2.0", "xgboost>=2.0", "scikit-learn>=1.3", "numpy>=1.24"]

[project.scripts]
autoresearch = "autoresearch.cli:main"

[project.urls]
Homepage = "https://github.com/anonymous/autoresearch"

[tool.setuptools]
packages = ["autoresearch"]
include-package-data = true

[tool.setuptools.package-data]
autoresearch = ["../templates/*.template", "../examples/**/*"]
